drf/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
django-hosts==6.0
django-subdomains==2.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
dnspython==2.7.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
PySocks==1.7.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
